% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{radiant}
\alias{radiant}
\title{Launch Radiant Lite in the default browser}
\usage{
radiant()
}
\description{
Launch Radiant Lite in the default browser
}
\details{
See \url{https://radiant-r.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
radiant.lite::radiant()
}
}
