% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/serenity.R
\docType{package}
\name{serenity}
\alias{serenity}
\alias{serenity-package}
\alias{serenity}
\title{serenity}
\usage{
serenity()
}
\description{
serenity

Launch Serenity in the default browser
}
\details{
See \url{https://serenity-r.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
serenity::serenity()
}
}
